backend/requirements.txt
anyio==3.6.1
asgiref==3.5.2
asn1crypto==1.5.1
boto3==1.24.19
botocore==1.27.19
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
ecdsa==0.17.0
fastapi==0.78.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
jmespath==1.0.1
mypy-extensions==0.4.3
mysql-connector==2.2.9
pathspec==0.9.0
platformdirs==2.5.2
pyasn1==0.4.8
pycodestyle==2.8.0
pydantic==1.9.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.27.1
rsa==4.8
s3transfer==0.6.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
bcrypt==3.2.2
